pyproject.toml
[tool.poetry]
name = "webscrapingutils"
version = "0.1.0"
description = "Utilitary functions to web scrap using proxies"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "4.12.3"
pytest-cov = "^5.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
